% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_importance.R
\name{plot_variable_importance}
\alias{plot_variable_importance}
\title{Variable importance of Bayesian Additive Regression Trees}
\usage{
plot_variable_importance(.model, confounders, out = "all")
}
\arguments{
\item{.model}{a model produced by bartCause::bartc(). Typically store$model_results}

\item{confounders}{matrix of confounders}
}
\value{
a list containing variable importance plot & ordered table of confounders by scaled importance
}
\description{
Fit single regression tree on bartc() icates to produce variable importance plot & table.
}
\author{
George Perrett, Joe Marlo
}
