% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_calculate.R
\name{roc_calculate}
\alias{roc_calculate}
\title{Calculate roc from formed input object}
\usage{
roc_calculate(tmp_res2, tmp_J, J_name = "None", mark_multiple = TRUE)
}
\arguments{
\item{tmp_res2}{a correlation object}

\item{tmp_J}{a formed database file}

\item{J_name}{label}

\item{mark_multiple}{boolean value representing database file of one category or multiple categories}
}
\value{
A list of the calculated roc in indicated database file (roc_qda,plotx,roc.info).
}
\description{
This function loads a correlation object and a formed database file as input, return a list of calculated roc.
}
